+++ include/Index.h
@@ -23,15 +23,15 @@ class Index {
             int read = 0;
         };
 
-        mSearchResult mSearch(double x);
+        mSearchResult mSearch(int x);
         tuple<int, int> insertB(DataFile::registry* newRegistry);
                 
     private:
         TreeFile* treeFile;
         DataFile* dataFile;
         
-        int linearSearch(vector<double> &K, double x, int n);
-        mSearchResult mSearch(TreeFile* treeFile, double x);
+        int linearSearch(vector<int> &K, int x, int n);
+        mSearchResult mSearch(TreeFile* treeFile, int x);
         tuple<int, int> insertB(DataFile::registry* newRegistry, TreeFile* treeFile, DataFile* dataFile);
 
 };
